requirements.txt
python-telegram-bot==21.7
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.5
pytest-asyncio==0.25.3
